requirements.txt
clickhouse_driver==0.2.4
DingtalkChatbot==1.5.3
loguru==0.5.3
google-cloud-bigquery==3.3.0